requirements.txt
pathway
groq
openai
sentence-transformers
transformers
tqdm
pandas
numpy
python-dotenv
scikit-learn
torch